refactor: consolidate duplicated SVG icons into shared libraries
Create icons.tsx (28 generic icons) and brand-icons.tsx (ShopifyIcon,
DataSourceIcon) as canonical sources. Replace local definitions in
presentation.tsx, panels.tsx, JobDetailPanel.tsx, LabelPreview.tsx,
and RecentSourcesModal.tsx with imports. Add re-exports for backward
compatibility where external consumers import from the original files.
