+++ frontend/src/components/sidebar/panels.tsx
@@ -16,107 +16,23 @@ import { getJobs, deleteJob, disconnectDataSource, importDataSource, uploadDataS
 import type { Job, JobSummary, DataSourceInfo, PlatformType } from '@/types/api';
 import { RecentSourcesModal } from '@/components/RecentSourcesModal';
 import { toDataSourceColumns } from '@/components/sidebar/dataSourceMappers';
+import {
+  ChevronIcon,
+  SearchIcon,
+  HardDriveIcon,
+  CloudIcon,
+  EyeIcon,
+  EyeOffIcon,
+  TrashIcon,
+  PrinterIcon,
+  HistoryIcon,
+} from '@/components/ui/icons';
+import { ShopifyIcon, DataSourceIcon } from '@/components/ui/brand-icons';
 
 /* Future: more external platforms like WooCommerce, SAP, Oracle */
 
-
-// Icons
-export function ChevronIcon({ direction, className }: { direction: 'left' | 'right'; className?: string }) {
-  return (
-    <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth={2} className={className}>
-      {direction === 'left' ? (
-        <path d="M15 18l-6-6 6-6" strokeLinecap="round" strokeLinejoin="round" />
-      ) : (
-        <path d="M9 18l6-6-6-6" strokeLinecap="round" strokeLinejoin="round" />
-      )}
-    </svg>
-  );
-}
-
-export function SearchIcon({ className }: { className?: string }) {
-  return (
-    <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth={1.5} className={className}>
-      <circle cx="11" cy="11" r="8" />
-      <path d="M21 21l-4.35-4.35" strokeLinecap="round" strokeLinejoin="round" />
-    </svg>
-  );
-}
-
-export function HardDriveIcon({ className }: { className?: string }) {
-  return (
-    <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth={1.5} className={className}>
-      <path d="M22 12H2" strokeLinecap="round" strokeLinejoin="round" />
-      <path d="M5.45 5.11L2 12v6a2 2 0 002 2h16a2 2 0 002-2v-6l-3.45-6.89A2 2 0 0016.76 4H7.24a2 2 0 00-1.79 1.11z" strokeLinecap="round" strokeLinejoin="round" />
-      <line x1="6" y1="16" x2="6.01" y2="16" strokeLinecap="round" strokeLinejoin="round" />
-      <line x1="10" y1="16" x2="10.01" y2="16" strokeLinecap="round" strokeLinejoin="round" />
-    </svg>
-  );
-}
-
-export function CloudIcon({ className }: { className?: string }) {
-  return (
-    <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth={1.5} className={className}>
-      <path d="M18 10h-1.26A8 8 0 109 20h9a5 5 0 000-10z" strokeLinecap="round" strokeLinejoin="round" />
-    </svg>
-  );
-}
-
-export function ShopifyIcon({ className }: { className?: string }) {
-  return (
-    <svg viewBox="0 0 24 24" fill="none" className={className}>
-      <text
-        x="12"
-        y="17"
-        textAnchor="middle"
-        fontFamily="system-ui, -apple-system, sans-serif"
-        fontSize="18"
-        fontWeight="700"
-        fill="currentColor"
-      >
-        S
-      </text>
-    </svg>
-  );
-}
-
-export function EyeIcon({ className }: { className?: string }) {
-  return (
-    <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth={1.5} className={className}>
-      <path d="M1 12s4-8 11-8 11 8 11 8-4 8-11 8-11-8-11-8z" strokeLinecap="round" strokeLinejoin="round" />
-      <circle cx="12" cy="12" r="3" />
-    </svg>
-  );
-}
-
-export function EyeOffIcon({ className }: { className?: string }) {
-  return (
-    <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth={1.5} className={className}>
-      <path d="M17.94 17.94A10.07 10.07 0 0112 20c-7 0-11-8-11-8a18.45 18.45 0 015.06-5.94M9.9 4.24A9.12 9.12 0 0112 4c7 0 11 8 11 8a18.5 18.5 0 01-2.16 3.19m-6.72-1.07a3 3 0 11-4.24-4.24" strokeLinecap="round" strokeLinejoin="round" />
-      <line x1="1" y1="1" x2="23" y2="23" strokeLinecap="round" strokeLinejoin="round" />
-    </svg>
-  );
-}
-
-export function TrashIcon({ className }: { className?: string }) {
-  return (
-    <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth={1.5} className={className}>
-      <polyline points="3 6 5 6 21 6" strokeLinecap="round" strokeLinejoin="round" />
-      <path d="M19 6v14a2 2 0 01-2 2H7a2 2 0 01-2-2V6m3 0V4a2 2 0 012-2h4a2 2 0 012 2v2" strokeLinecap="round" strokeLinejoin="round" />
-      <line x1="10" y1="11" x2="10" y2="17" strokeLinecap="round" strokeLinejoin="round" />
-      <line x1="14" y1="11" x2="14" y2="17" strokeLinecap="round" strokeLinejoin="round" />
-    </svg>
-  );
-}
-
-export function PrinterIcon({ className }: { className?: string }) {
-  return (
-    <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth={1.5} className={className}>
-      <polyline points="6 9 6 2 18 2 18 9" strokeLinecap="round" strokeLinejoin="round" />
-      <path d="M6 18H4a2 2 0 01-2-2v-5a2 2 0 012-2h16a2 2 0 012 2v5a2 2 0 01-2 2h-2" strokeLinecap="round" strokeLinejoin="round" />
-      <rect x="6" y="14" width="12" height="8" strokeLinecap="round" strokeLinejoin="round" />
-    </svg>
-  );
-}
+// Re-export icons for backward compatibility (Sidebar.tsx imports from here).
+export { ChevronIcon, SearchIcon, HardDriveIcon, CloudIcon, ShopifyIcon, EyeIcon, EyeOffIcon, TrashIcon, PrinterIcon, DataSourceIcon, HistoryIcon };
 
 // Status badge component
 export function StatusBadge({ status }: { status: string }) {
@@ -920,27 +836,4 @@ export function JobHistorySection({
   );
 }
 
-// Mini icons for collapsed state
-export function DataSourceIcon({ className, connected }: { className?: string; connected?: boolean }) {
-  return (
-    <div className="relative">
-      <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth={1.5} className={className}>
-        <ellipse cx="12" cy="6" rx="8" ry="3" />
-        <path d="M4 6v6c0 1.657 3.582 3 8 3s8-1.343 8-3V6" />
-        <path d="M4 12v6c0 1.657 3.582 3 8 3s8-1.343 8-3v-6" />
-      </svg>
-      {connected && (
-        <span className="absolute -top-0.5 -right-0.5 w-2 h-2 bg-success rounded-full" />
-      )}
-    </div>
-  );
-}
-
-export function HistoryIcon({ className }: { className?: string }) {
-  return (
-    <svg viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth={1.5} className={className}>
-      <circle cx="12" cy="12" r="10" />
-      <polyline points="12,6 12,12 16,14" strokeLinecap="round" strokeLinejoin="round" />
-    </svg>
-  );
-}
+// Mini icons for collapsed state — imported from shared libraries, re-exported above.
